+++ consistent-sampling/src/main/java/io/opentelemetry/contrib/samplers/ConsistentAlwaysOffSampler.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.contrib.samplers;
 
-import io.opentelemetry.contrib.state.OtelTraceState;
 import javax.annotation.concurrent.Immutable;
 
 @Immutable
