+++ consistent-sampling/src/main/java/io/opentelemetry/contrib/samplers/ConsistentProbabilityBasedSampler.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.contrib.samplers;
 
-import io.opentelemetry.contrib.util.RandomGenerator;
 import javax.annotation.concurrent.Immutable;
 
 /** A consistent sampler that samples with a fixed probability. */
