+++ consistent-sampling/src/test/java/io/opentelemetry/contrib/samplers/OtelTraceStateTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.state;
+package io.opentelemetry.contrib.samplers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
